Conflicts:
	android/jni/src/Android.mk
	src/gfx/Pov.cpp
	src/main.cpp
	src/sim/TouchObserverController.cpp

Former-commit-id: 111d73bfaae8c5e5e4e1580bf39836269f43bca4

+++ src/gfx/Scene.cpp
@@ -31,7 +31,7 @@ Scene::Scene(Engine & engine)
 {
 	auto frustum = pov.GetFrustum();
 
-	frustum.fov = static_cast<double>(camera_fov);
+	frustum.fov = camera_fov;
 	frustum.depth_range[0] = frustum.depth_range[1] = -1;
 	
 	pov.SetFrustum(frustum);
